Add files via upload
This Java program performs a linear search. It asks the user to enter an array of numbers and then searches for a specific number in that array. If the number is found, the program displays its position; if not, it informs the user that the number is not in the array.